engine: fail loudly on LTM ceteris-paribus parse failure
The LTM ceteris-paribus partial-equation builders
(build_partial_equation_shaped / _with_live_ref and
subscript_idents_at_element) silently fell back to the lowercased input
text when Expr0::new returned Err or Ok(None). With no PREVIOUS() wrapping
applied, that "partial" is identical to the target's full equation, so the
link-score numerator (partial - PREVIOUS(target)) equals the denominator
(target - PREVIOUS(target)) and the score magnitude collapses to a constant
|Δz/Δz| = 1. Worse, that wrong equation compiles cleanly, so
model_ltm_fragment_diagnostics (which only catches *compile* failures) never
sees it -- a hidden attribution error rather than an explicit failure
(GH #311).

The fix makes the builders return a structured PartialEquationError carrying
the original equation text, and threads it through every link-score
generator (generate_link_score_equation_for_link, the aux-to-aux /
stock-to-flow / arrayed-per-element generators, and the agg-half emitters).
The db-bearing callers (link_score_equation_text,
link_score_equation_text_shaped, and the scalar->arrayed / agg->target
emitters in db/ltm) convert the error into a Warning naming the variable and
the offending equation text, and skip emitting that link-score variable --
the established "loud failure" pattern (cf.
emit_unscoreable_disjoint_edge_warning). Chose option (a) loud-skip over a
semantics-preserving fallback because any silent degradation here (even a
safe-zero partial) remains hidden, which is exactly the property the issue
calls out as dangerous.

The legitimate "parsed fine, no other deps to wrap" case (a constant, or the
live source alone) is unchanged: it still returns Ok with the re-printed
text, distinguishing it from the parse-failure path the bug blurred. The
parse failure is effectively unreachable through the production salsa path
(the equation text is always a print_eqn re-print, and an empty equation is
rejected as an EmptyEquation Error upstream before LTM augmentation runs), so
this is defense-in-depth for the unanticipated case -- which is why the
silent semantics-changing fallback was so hazardous. Builder-level failure is
tested directly (build_partial_equation_shaped_parse_error_is_err /
_empty_equation_is_err / _no_deps_to_wrap_is_ok,
subscript_idents_at_element_parse_error_is_err), the generator chain
end-to-end (generate_link_score_equation_for_link_empty_target_is_err /
_normal_target_is_ok), the db warning message contract
(test_ltm_partial_equation_warning_message_contract), and the actual salsa
accumulation through a db-bearing emitter
(test_ltm_partial_equation_warning_accumulates_via_salsa, which drives the
real .accumulate(db) path the six call sites share since the error-producing
production route is unreachable). The test module is split into the sibling
ltm_augment_tests.rs to keep ltm_augment.rs under the per-file line cap.
